+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/repository/jwt/JwtProperties.java
@@ -1,4 +1,4 @@
-package kr.modernworld.modernworldv2.user.infrastructure.repository;
+package kr.modernworld.modernworldv2.user.infrastructure.repository.jwt;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
@@ -6,8 +6,8 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 public record JwtProperties(
     String accessSecret,
     String refreshSecret,
-    long accessExpiration,
-    long refreshExpiration,
+    Long accessExpiration,
+    Long refreshExpiration,
     String issuer
 ) {
